Authentifizierungsabfragen hinzugefügt

+++ GoAppServer/src/kit/edu/pse/goapp/server/servlets/GroupUserManagementServlet.java
@@ -13,6 +13,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import kit.edu.pse.goapp.server.converter.daos.GroupMemberDaoConverter;
 import kit.edu.pse.goapp.server.converter.objects.ObjectConverter;
@@ -20,6 +21,8 @@ import kit.edu.pse.goapp.server.daos.GroupDAO;
 import kit.edu.pse.goapp.server.daos.GroupDaoImpl;
 import kit.edu.pse.goapp.server.daos.GroupMemberDAO;
 import kit.edu.pse.goapp.server.daos.GroupMemberDaoImpl;
+import kit.edu.pse.goapp.server.daos.UserDAO;
+import kit.edu.pse.goapp.server.daos.UserDaoImpl;
 import kit.edu.pse.goapp.server.datamodels.Group;
 import kit.edu.pse.goapp.server.datamodels.User;
 import kit.edu.pse.goapp.server.exceptions.CustomServerException;
@@ -47,6 +50,7 @@ public class GroupUserManagementServlet extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 		try {
+			int userId = authenticateUser(request);
 			String groupId = request.getParameter("groupId");
 			GroupMemberDAO dao = new GroupMemberDaoImpl();
 			try {
@@ -55,6 +59,9 @@ public class GroupUserManagementServlet extends HttpServlet {
 				throw new CustomServerException("The GroupID from the JSON string isn't correct!",
 						HttpServletResponse.SC_BAD_REQUEST);
 			}
+			User user = createUserWithDao(userId);
+			Group groupCheckingMember = createGroupWithDao(dao.getGroupId());
+			groupCheckingMember.isMember(user);
 			if (dao != null) {
 				List<User> groupMembers = dao.getAllMembers();
 				response.getWriter().write(new ObjectConverter<List<User>>().serialize(groupMembers,
@@ -76,18 +83,17 @@ public class GroupUserManagementServlet extends HttpServlet {
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 		try {
+			int userId = authenticateUser(request);
 			String jsonString = request.getReader().readLine();
 			GroupMemberDAO groupMemberDao = new GroupMemberDaoConverter().parse(jsonString);
 			// every group member is an admin automatically (priority A
 			// requirement)
 			groupMemberDao.setAdmin(true);
+			checkIfUserIsMemberAndAdmin(userId, groupMemberDao.getGroupId());
 			if (groupMemberDao != null) {
 				groupMemberDao.addMember();
 			}
-
-			GroupDAO groupDao = new GroupDaoImpl();
-			groupDao.setGroupId(groupMemberDao.getGroupId());
-			Group group = groupDao.getGroupByID();
+			Group group = createGroupWithDao(groupMemberDao.getGroupId());
 			response.getWriter().write(new ObjectConverter<Group>().serialize(group, Group.class));
 		} catch (CustomServerException e) {
 			response.setStatus(e.getStatusCode());
@@ -104,14 +110,14 @@ public class GroupUserManagementServlet extends HttpServlet {
 	protected void doPut(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 		try {
+			int userId = authenticateUser(request);
 			String jsonString = request.getReader().readLine();
 			GroupMemberDAO groupMemberdao = new GroupMemberDaoConverter().parse(jsonString);
+			checkIfUserIsMemberAndAdmin(userId, groupMemberdao.getGroupId());
 			if (groupMemberdao != null) {
 				groupMemberdao.updateMember();
 			}
-			GroupDAO groupDao = new GroupDaoImpl();
-			groupDao.setGroupId(groupMemberdao.getGroupId());
-			Group group = groupDao.getGroupByID();
+			Group group = createGroupWithDao(groupMemberdao.getGroupId());
 			response.getWriter().write(new ObjectConverter<Group>().serialize(group, Group.class));
 		} catch (CustomServerException e) {
 			response.setStatus(e.getStatusCode());
@@ -128,6 +134,7 @@ public class GroupUserManagementServlet extends HttpServlet {
 	protected void doDelete(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 		try {
+			int myUserId = authenticateUser(request);
 			String groupId = request.getParameter("groupId");
 			String userId = request.getParameter("userId");
 			GroupMemberDAO dao = new GroupMemberDaoImpl();
@@ -143,6 +150,7 @@ public class GroupUserManagementServlet extends HttpServlet {
 				throw new CustomServerException("The UserID from the JSON string isn't correct!",
 						HttpServletResponse.SC_BAD_REQUEST);
 			}
+			checkIfUserIsMemberAndAdmin(myUserId, dao.getGroupId());
 			if (dao != null) {
 				dao.deleteMember();
 			}
@@ -152,5 +160,36 @@ public class GroupUserManagementServlet extends HttpServlet {
 			response.getWriter().write(e.getMessage());
 		}
 
+	}
+
+	private int authenticateUser(HttpServletRequest request) throws CustomServerException {
+		HttpSession session = request.getSession(true);
+
+		int userId = 1;// (int) session.getAttribute("userId");
+		if (userId <= 0) {
+			throw new CustomServerException("This user is unauthorized!", HttpServletResponse.SC_UNAUTHORIZED);
+		}
+		return userId;
+	}
+
+	private Group createGroupWithDao(int groupId) throws IOException, CustomServerException {
+		GroupDAO dao = new GroupDaoImpl();
+		dao.setGroupId(groupId);
+		Group group = dao.getGroupByID();
+		return group;
+	}
+
+	private User createUserWithDao(int userId) throws IOException, CustomServerException {
+		UserDAO userDao = new UserDaoImpl();
+		userDao.setUserId(userId);
+		User user = userDao.getUserByID();
+		return user;
+	}
+
+	private void checkIfUserIsMemberAndAdmin(int userId, int groupId) throws IOException, CustomServerException {
+		User user = createUserWithDao(userId);
+		Group groupCheckingAuthorization = createGroupWithDao(groupId);
+		groupCheckingAuthorization.isMember(user);
+		groupCheckingAuthorization.isAdmin(user);
 	}
 }
