+++ GoAppServer/src/kit/edu/pse/goapp/server/servlets/UserServlet.java
@@ -12,7 +12,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.xml.ws.http.HTTPException;
+import javax.servlet.http.HttpSession;
 
 import kit.edu.pse.goapp.server.converter.daos.UserDaoConverter;
 import kit.edu.pse.goapp.server.converter.objects.ObjectConverter;
@@ -72,18 +72,22 @@ public class UserServlet extends HttpServlet {
 	@Override
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-
-		String googleId = request.getParameter("googleId");
-		String jsonString = request.getReader().readLine();
-		UserDAO dao = new UserDaoConverter().parse(jsonString);
-		// every user has notifications enabled by default
-		dao.setNotificationEnabled(true);
-		dao.setGoogleId(googleId);
-		if (dao != null) {
-			dao.addUser();
+		try {
+			String googleId = request.getParameter("googleId");
+			String jsonString = request.getReader().readLine();
+			UserDAO dao = new UserDaoConverter().parse(jsonString);
+			// every user has notifications enabled by default
+			dao.setNotificationEnabled(true);
+			dao.setGoogleId(googleId);
+			if (dao != null) {
+				dao.addUser();
+			}
+			User user = dao.getUserByID();
+			response.getWriter().write(new ObjectConverter<User>().serialize(user, User.class));
+		} catch (CustomServerException e) {
+			response.setStatus(e.getStatusCode());
+			response.getWriter().write(e.toString());
 		}
-		User user = dao.getUserByID();
-		response.getWriter().write(new ObjectConverter<User>().serialize(user, User.class));
 
 	}
 
@@ -95,15 +99,23 @@ public class UserServlet extends HttpServlet {
 	@Override
 	protected void doPut(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-		String jsonString = request.getReader().readLine();
-		UserDAO dao = new UserDaoConverter().parse(jsonString);
-		if (dao != null) {
-			dao.updateUser();
-		} else {
-			throw new HTTPException(HttpServletResponse.SC_BAD_REQUEST);
+		try {
+			int userId = authenticateUser(request);
+			String jsonString = request.getReader().readLine();
+			UserDAO dao = new UserDaoConverter().parse(jsonString);
+			if (userId != dao.getUserId()) {
+				throw new CustomServerException("The UserID from the JSON string isn't the same as the actual user!",
+						HttpServletResponse.SC_BAD_REQUEST);
+			}
+			if (dao != null) {
+				dao.updateUser();
+			}
+			User user = dao.getUserByID();
+			response.getWriter().write(new ObjectConverter<User>().serialize(user, User.class));
+		} catch (CustomServerException e) {
+			response.setStatus(e.getStatusCode());
+			response.getWriter().write(e.toString());
 		}
-		User user = dao.getUserByID();
-		response.getWriter().write(new ObjectConverter<User>().serialize(user, User.class));
 	}
 
 	/**
@@ -114,13 +126,39 @@ public class UserServlet extends HttpServlet {
 	@Override
 	protected void doDelete(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-		String userId = request.getParameter("userId");
-		UserDAO dao = new UserDaoImpl();
-		dao.setUserId(Integer.parseInt(userId));
-		if (dao != null) {
-			dao.deleteUser();
+		try {
+			int authenticatingUserId = authenticateUser(request);
+			String userId = request.getParameter("userId");
+			UserDAO dao = new UserDaoImpl();
+			try {
+				dao.setUserId(Integer.parseInt(userId));
+			} catch (Exception e) {
+				throw new CustomServerException("The UserID from the JSON string isn't correct!",
+						HttpServletResponse.SC_BAD_REQUEST);
+			}
+			if (authenticatingUserId != Integer.parseInt(userId)) {
+				throw new CustomServerException("The UserID from the JSON string isn't the same as the actual user!",
+						HttpServletResponse.SC_BAD_REQUEST);
+			}
+			if (dao != null) {
+				dao.deleteUser();
+			}
+			response.setStatus(HttpServletResponse.SC_OK);
+
+		} catch (CustomServerException e) {
+			response.setStatus(e.getStatusCode());
+			response.getWriter().write(e.toString());
 		}
-		response.setStatus(HttpServletResponse.SC_OK);
+	}
+
+	private int authenticateUser(HttpServletRequest request) throws CustomServerException {
+		HttpSession session = request.getSession(true);
+
+		int userId = 1;// (int) session.getAttribute("userId");
+		if (userId <= 0) {
+			throw new CustomServerException("This user is unauthorized!", HttpServletResponse.SC_UNAUTHORIZED);
+		}
+		return userId;
 	}
 
 }
